test(runtime): split review regression coverage
